% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mystats}
\alias{mystats}
\title{mystats}
\usage{
mystats(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector for which the statistics are to be calculated.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
A list of statistics including mean, median, standard deviation, median absolute deviation, variance, minimum, maximum, and various quantiles.
}
\description{
This function calculates various statistics for a given numeric vector.
}
\examples{
mystats(c(1, 2, 3, 4, 5)) # returns a list of statistics
mystats(c(NA, 2, 3, 4, 5), na.rm = TRUE) # returns a list of statistics with NA removed
}
