% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onenesstwoness-package.R
\docType{package}
\name{onenesstwoness-package}
\alias{onenesstwoness}
\alias{onenesstwoness-package}
\title{onenesstwoness - predictor of HRD phenotype}
\description{
This package provides a random forest classifier for BRCA1/BRCA2 deletion phenotype classification.
}
\author{
\strong{Maintainer}: Johnathan Rafailov \email{jrafailov@nygenome.org}

Authors:
\itemize{
  \item Zi-Ning Choo \email{zic2001@med.cornell.edu}
  \item Kevin Hadi \email{khadi@nygenome.org}
}

}
