% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.pp}
\alias{predict.pp}
\title{predict.pp}
\usage{
\method{predict}{pp}(pp.res, newdat, apply.prefun = TRUE)
}
\arguments{
\item{pp.res}{A list containing the preprocessed data.}

\item{newdat}{A data frame containing the new data to be predicted.}

\item{apply.prefun}{A logical value indicating whether the pre-processing function should be applied.}
}
\value{
A list containing the preprocessed data, the new data, the pre-processing function, the variables, the method, and the original new data.
}
\description{
This function predicts using preprocessed data.
}
