% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_B1_2}
\alias{predict_B1_2}
\title{predict_B1_2}
\usage{
predict_B1_2(
  complex,
  homeology = NULL,
  homeology_stats = NULL,
  hrdetect_results = NULL,
  overwrite = TRUE,
  snv = NULL,
  indel = NULL,
  jabba = NULL,
  sv = NULL,
  mask = NULL,
  hets = NULL,
  genome = "~/DB/GATK/human_g1k_v37.fasta",
  width = 200,
  pad = 20,
  thresh = 1,
  stride = 8,
  ref = NULL,
  model = system.file("model", "stash.retrained.model.rds", package = "onenesstwoness"),
  outdir = "./",
  cores = 4,
  save = TRUE
)
}
\arguments{
\item{complex}{Path to the complex events gGraph (required)}

\item{homeology}{Path to the homeology file (optional)}

\item{homeology_stats}{Path to the homeology stats file (optional)}

\item{hrdetect_results}{Path to the HRDetect results file (optional)}

\item{overwrite}{Logical indicating whether to overwrite existing files (default = TRUE)}

\item{snv}{Path to the SNV file (optional)}

\item{width}{Window size for homeology analysis (default = 200)}

\item{pad}{Padding size for homeology analysis (default = 20)}

\item{thresh}{Threshold for homeology analysis (default = 1)}

\item{stride}{Stride size for homeology analysis (default = 8)}

\item{model}{Path to the model file(default = system.file("data/model", "stash.retrained.model.rds", package = "onenesstwoness"))}

\item{cores}{Number of cores to use (default = 4)}

\item{save}{Logical indicating whether to save the results}
}
\value{
A list containing the following elements:
\itemize{
\item{expl_variables}{A data frame containing the explanatory variables}
\item{ot_scores}{A data frame containing the Oneness and Twoness scores}
}
}
\description{
function to predict the B1+2 score
requires Complex Events gGraph, Homeology, Homeology Stats, HRDetect results
if homeology, homeology_stats are not provided, the function will run homeology
if hrdetect_results is not provided, the function will run HRDetect
}
