% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_hrd}
\alias{predict_hrd}
\title{predict_hrd}
\usage{
predict_hrd(
  complex,
  homeology,
  homeology_stats,
  hrdetect_results,
  model = system.file("data/model", "stash.retrained.model.rds", package =
    "onenesstwoness"),
  libdir
)
}
\arguments{
\item{complex}{Path to the complex file}

\item{homeology}{Path to the homeology file}

\item{homeology_stats}{Path to the homeology stats file}

\item{hrdetect_results}{Path to the HRDetect results file}

\item{model}{Path to the model file}

\item{libdir}{Path to the library directory}
}
\value{
A list containing the following elements:
\itemize{
\item{expl_variables}{A data frame containing the explanatory variables}
\item{ot_scores}{A data frame containing the Oneness and Twoness scores}
}
}
\description{
function to predict the HRD score
}
