% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{run_hrdetect}
\alias{run_hrdetect}
\title{run_hrdetect}
\usage{
run_hrdetect(
  snv,
  hets,
  jabba = "/dev/null",
  indel = snv,
  sv = "/dev/null",
  mask = system.file("extdata", "hg19_mask.rds", package = "onenesstwoness"),
  genome,
  ref = "hg19",
  save = TRUE,
  outdir = "./"
)
}
\arguments{
\item{snv}{Path to the SNV file}

\item{hets}{Path to the hets file}

\item{jabba}{Jabba rds file with $asegstats}

\item{indel}{Path to the indel file if not present in SNV}

\item{sv}{Path to the SV file (vcf/rds)}

\item{mask}{Path to the mask file}

\item{genome}{Path to the reference genome (fasta)}

\item{ref}{enum of hg19/hg38/mm10/canFam3}

\item{save}{Logical indicating whether to save the results}

\item{outdir}{Path to the output directory}
}
\description{
function to run HRDetect
}
