% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onenesstwoness-package.R
\docType{package}
\name{onenesstwoness-package}
\alias{-package}
\alias{onenesstwoness-package}
\title{onenesstwoness - predictor of HRD phenotype}
\description{
This package provides tools for predicting HRD (Homologous Recombination Deficiency) phenotype 
using various genomic and statistical methods. It integrates multiple libraries and utilities 
to facilitate genomic data analysis and modeling.
}
